# src/main/java/de/starwit/BiastestsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package de.starwit;

import de.starwit.aicockpit.BiasTest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-04-01T15:47:00.058415289+02:00[Europe/Berlin]")
@Validated
@Tag(name = "biastests", description = "the biastests API")
public interface BiastestsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /biastests
     * create a bias test
     *
     * @param biasTest Bias test to be created (required)
     * @return create new bias test (status code 200)
     */
    @Operation(
        operationId = "createBiasTest",
        description = "create a bias test",
        responses = {
            @ApiResponse(responseCode = "200", description = "create new bias test")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/biastests",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> createBiasTest(
        @Parameter(name = "BiasTest", description = "Bias test to be created", required = true) @Valid @RequestBody BiasTest biasTest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /biastests
     * get all biastests for this system
     *
     * @return list all biastests in this system (status code 200)
     */
    @Operation(
        operationId = "getAllBiasTests",
        description = "get all biastests for this system",
        responses = {
            @ApiResponse(responseCode = "200", description = "list all biastests in this system", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = BiasTest.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/biastests",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<BiasTest>> getAllBiasTests(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"name\" : \"name\", \"id\" : 0 }, { \"name\" : \"name\", \"id\" : 0 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /biastests/{id}
     * update a bias test
     *
     * @param id Numeric ID of bias test to be updated (required)
     * @param biasTest Bias test to be updated (required)
     * @return update a bias test (status code 200)
     */
    @Operation(
        operationId = "updateBiasTest",
        description = "update a bias test",
        responses = {
            @ApiResponse(responseCode = "200", description = "update a bias test")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/biastests/{id}",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> updateBiasTest(
        @Parameter(name = "id", description = "Numeric ID of bias test to be updated", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "BiasTest", description = "Bias test to be updated", required = true) @Valid @RequestBody BiasTest biasTest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}